backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
google-api-python-client==2.108.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
django-redis==5.2.0
redis==5.0.1
beautifulsoup4==4.12.2
requests==2.31.0

